refactor: `config.py` moved to `driver` dir
